+++ ConsoleExample/Examples/ExampleUtils.cs
@@ -11,12 +11,12 @@ namespace ConsoleExample.Examples
         /// <summary>
         /// Команда отработает для бота с botId 0.
         /// Команда отработает при написание в чат "Awaiter message".
-        /// Сначало будет отправлено сообщение о генерации данных, после двух секунд старое сообщение будет удалено и сразу появится новое. 
+        /// Сначало будет отправлено сообщение 'Обработка данных...', после двух секунд старое сообщение будет удалено и сразу появится новое. 
         /// </summary>
         [ReplyMenuHandler("Awaiter message")]
-        public static async Task AdminExample(ITelegramBotClient botClient, Update update)
+        public static async Task AwaiterExample (ITelegramBotClient botClient, Update update)
         {
-            using(var messageAwaiter = new MessageAwaiter(botClient, update.GetChatId()))
+            using(var messageAwaiter = new MessageAwaiter(botClient, update.GetChatId(), "Обработка данных..."))
             {
                 // Симуляция тяжелой операции.
                 await Task.Delay(2000);