+++ PRTelegramBot/Models/TCommands/TCommandBase.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using PRTelegramBot.Models.Enums;
+using System.Text.Json.Serialization;
 
 namespace PRTelegramBot.Models.CallbackCommands
 {
@@ -12,8 +13,14 @@ namespace PRTelegramBot.Models.CallbackCommands
         /// <summary>
         /// Предыдущая команда.
         /// </summary>
-        [JsonPropertyName("lc")]
+        [JsonPropertyName("l")]
         public int LastCommand { get; set; }
+
+        /// <summary>
+        /// Действие с предыдущим сообщением.
+        /// </summary>
+        [JsonPropertyName("a")]
+        public int ActionWithLastMessage { get; set; }
 
         #endregion
 
@@ -29,6 +36,15 @@ namespace PRTelegramBot.Models.CallbackCommands
             return (TEnum)Enum.ToObject(typeof(TEnum), LastCommand);
         }
 
+        /// <summary>
+        /// Действие с последним сообщением.
+        /// </summary>
+        /// <returns>Enum что сделать с последним сообщением.</returns>
+        public ActionWithLastMessage GetActionWithLastMessage()
+        {
+            return (ActionWithLastMessage)Enum.ToObject(typeof(ActionWithLastMessage), ActionWithLastMessage);
+        }
+
         #endregion
 
         #region Конструкторы
@@ -37,9 +53,7 @@ namespace PRTelegramBot.Models.CallbackCommands
         /// Конструктор.
         /// </summary>
         public TCommandBase()
-        {
-            LastCommand = 0;
-        }
+            : this(0) { }
 
         /// <summary>
         /// Конструктор.
@@ -48,6 +62,27 @@ namespace PRTelegramBot.Models.CallbackCommands
         public TCommandBase(int command)
         {
             LastCommand = command;
+        }
+
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="command">Команда.</param>
+        /// <param name="action">Действие с последним сообщением.</param>
+        public TCommandBase(int command, ActionWithLastMessage action)
+            : this (action)
+        {
+            LastCommand = command;
+        }
+
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="action">Действие с последним сообщением.</param>
+        public TCommandBase(ActionWithLastMessage action)
+            : this(0)
+        {
+            ActionWithLastMessage = (int)action;
         }
 
         #endregion